requirements.txt
SpeechRecognition
pyttsx3
pyaudio
keyboard 
requests 
winsdk 
python-dotenv 